Make delete applicant also delete interview
